+++ cmd/verkle/verkletrie/workers.go
@@ -22,8 +22,8 @@ import (
 	libcommon "github.com/erigontech/erigon-lib/common"
 	"github.com/holiman/uint256"
 
+	"github.com/erigontech/erigon-lib/trie/vtree"
 	"github.com/erigontech/erigon-lib/types/accounts"
-	"github.com/erigontech/erigon/turbo/trie/vtree"
 )
 
 type regeneratePedersenAccountsJob struct {
