+++ erigon-lib/trie/encoding.go
