+++ erigon-lib/trie/gen_struct_step.go
