+++ erigon-lib/trie/trie_root.go
@@ -126,7 +126,6 @@ type RootHashAggregator struct {
 	curr           bytes.Buffer   // Current key for the structure generation algorithm, as well as the input tape for the hash builder
 	succ           bytes.Buffer
 	currAccK       []byte
-	value          []byte // Current value to be used as the value tape for the hash builder
 	hadTreeAcc     bool
 	groups         []uint16 // `groups` parameter is the map of the stack. each element of the `groups` slice is a bitmask, one bit per element currently on the stack. See `GenStructStep` docs
 	hasTree        []uint16
