+++ erigon-lib/trie/vtree/verkle_utils.go
