+++ eth/stagedsync/stage_commit_rebuild.go
@@ -19,6 +19,7 @@ package stagedsync
 import (
 	"context"
 	"errors"
+
 	"github.com/erigontech/erigon/turbo/snapshotsync/freezeblocks"
 	"github.com/erigontech/erigon/turbo/stages/headerdownload"
 
@@ -28,7 +29,7 @@ import (
 	libcommon "github.com/erigontech/erigon-lib/common"
 	"github.com/erigontech/erigon-lib/kv"
 	"github.com/erigontech/erigon-lib/state"
-	"github.com/erigontech/erigon/turbo/trie"
+	"github.com/erigontech/erigon-lib/trie"
 )
 
 type TrieCfg struct {
